create a new module for partners defect
